+++ libs/angular-components/community/components/form/form-field/form-field.component.ts
@@ -1,0 +1,18 @@
+import {
+  ChangeDetectionStrategy,
+  Component,
+  ViewEncapsulation,
+} from "@angular/core";
+
+@Component({
+  selector: "tedi-form-field",
+  standalone: true,
+  template: "<ng-content />",
+  styleUrl: "./form-field.component.scss",
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  host: {
+    "[class.tedi-form-field]": "true",
+  },
+})
+export class FormFieldComponent {}